mic_test_crm/requirements.txt
drf-yasg>=1.21.8
mysqlclient>=2.2.7
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.4.0